test/web-platform-tests/to-upstream/html/semantics/forms/the-button-element/button-click-submits.html

<!DOCTYPE html>
<meta charset="utf-8">
<title>Clicking a button should submit the form</title>
<link rel="author" title="Domenic Denicola" href="mailto:d@domenic.me">
<link rel="help" href="https://html.spec.whatwg.org/multipage/#attr-button-type-submit-state">
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>

<script>
"use strict";

async_test(t => {

  const form = document.createElement("form");
  const button = document.createElement("button");
  form.appendChild(button);

  form.addEventListener("submit", t.step_func_done(ev => {
    assert_equals(ev.target, form);
  }));

  button.click();

}, "clicking a button with .click() should trigger a submit");

async_test(t => {

  const form = document.createElement("form");
  const button = document.createElement("button");
  form.appendChild(button);

  form.addEventListener("submit", t.step_func_done(ev => {
    assert_equals(ev.target, form);
  }));

  const e = new MouseEvent("click");
  button.dispatchEvent(e);

}, "clicking a button by dispatching an event should trigger a submit");

async_test(t => {

  const form = document.createElement("form");
  const button = document.createElement("button");
  form.appendChild(button);

  form.addEventListener("submit", t.unreached_func("Form should not be submitted"));
  button.addEventListener("click", t.step_func(ev => {
    ev.preventDefault();
    t.step_timeout(() => t.done(), 500);
  }));
  button.click();

}, "clicking a button that prevents the event's default should not trigger a submit");

async_test(t => {

  const form = document.createElement("form");
  const button = document.createElement("button");
  button.setAttribute("disabled", "");
  form.appendChild(button);

  form.addEventListener("submit", t.unreached_func("Form should not be submitted"));

  button.click();
  t.step_timeout(() => t.done(), 500);

}, "clicking a disabled button (via disabled attribute) should not trigger submit");

async_test(t => {

  const form = document.createElement("form");
  form.innerHTML = `<fieldset disabled><button>hello</button></fieldset>`;
  const button = form.querySelector("button");

  form.addEventListener("submit", t.unreached_func("Form should not be submitted"));

  button.click();
  t.step_timeout(() => t.done(), 500);

}, "clicking a disabled button (via ancestor fieldset) should not trigger submit");

test(t => {

  const form = document.createElement("form");
  form.innerHTML = `<fieldset disabled><legend><button>hello</button></legend></fieldset>`;
  const button = form.querySelector("button");

  form.addEventListener("submit", t.step_func_done(ev => {
    assert_equals(ev.target, form);
  }));

  button.click();

}, "clicking a button inside a disabled fieldset's legend *should* trigger submit");

</script>
